Convert the entire library into header only
Header only libraries are much easier to embed to projects, so convert
the entire library into such.
