+++ CMakeLists.txt
@@ -1,26 +1,75 @@
-CMAKE_MINIMUM_REQUIRED(VERSION 2.8.7)
-PROJECT(peelocpp_chrono C CXX)
+CMAKE_MINIMUM_REQUIRED(VERSION 3.12)
 
-INCLUDE(CheckCXXCompilerFlag)
+PROJECT(
+  PeeloChrono
+  VERSION 0.2.0
+  DESCRIPTION "Header only minimal C++ Gregorian calendar implementation."
+  HOMEPAGE_URL "https://github.com/peelonet/peelo-chrono"
+  LANGUAGES CXX
+)
 
-CHECK_CXX_COMPILER_FLAG("-std=c++11" COMPILER_SUPPORTS_CXX11)
-CHECK_CXX_COMPILER_FLAG("-std=c++0x" COMPILER_SUPPORTS_CXX0X)
+INCLUDE(GNUInstallDirs)
 
-IF(COMPILER_SUPPORTS_CXX11)
-  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
-ELSEIF(COMPILER_SUPPORTS_CXX0X)
-  SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++0x")
-ELSE()
-  MESSAGE(FATAL_ERROR "Compiler ${CMAKE_CXX_COMPILER} has no C++11 support.")
-ENDIF()
+ADD_LIBRARY(${PROJECT_NAME} INTERFACE)
 
-SET(EXECUTABLE_OUTPUT_PATH "${CMAKE_BINARY_DIR}/bin")
-SET(LIBRARY_OUTPUT_PATH "${CMAKE_BINARY_DIR}/lib")
+TARGET_INCLUDE_DIRECTORIES(
+  ${PROJECT_NAME}
+  INTERFACE
+    $<BUILD_INTERFACE:${${PROJECT_NAME}_SOURCE_DIR}/include>
+    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
+)
 
-INCLUDE_DIRECTORIES("${CMAKE_CURRENT_SOURCE_DIR}/include")
-LINK_DIRECTORIES(${LIBRARY_OUTPUT_PATH})
+TARGET_COMPILE_FEATURES(
+  ${PROJECT_NAME}
+  INTERFACE
+    cxx_std_17
+)
 
-ADD_SUBDIRECTORY(src)
+INSTALL(
+  TARGETS
+    ${PROJECT_NAME}
+  EXPORT
+    ${PROJECT_NAME}_Targets
+  ARCHIVE DESTINATION
+    ${CMAKE_INSTALL_LIBDIR}
+  LIBRARY DESTINATION
+    ${CMAKE_INSTALL_LIBDIR}
+  RUNTIME DESTINATION
+    ${CMAKE_INSTALL_BINDIR}
+)
+
+INCLUDE(CMakePackageConfigHelpers)
+WRITE_BASIC_PACKAGE_VERSION_FILE(
+  "${PROJECT_NAME}ConfigVersion.cmake"
+  VERSION
+    ${PROJECT_VERSION}
+  COMPATIBILITY
+    SameMajorVersion
+)
+CONFIGURE_PACKAGE_CONFIG_FILE(
+  "${PROJECT_SOURCE_DIR}/cmake/${PROJECT_NAME}Config.cmake.in"
+  "${PROJECT_BINARY_DIR}/${PROJECT_NAME}Config.cmake"
+  INSTALL_DESTINATION
+    ${CMAKE_INSTALL_DATAROOTDIR}/${PROJECT_NAME}/cmake
+)
+
+INSTALL(
+  EXPORT
+    ${PROJECT_NAME}_Targets
+  FILE
+    ${PROJECT_NAME}Targets.cmake
+  NAMESPACE
+    ${PROJECT_NAME}::
+  DESTINATION
+    ${CMAKE_INSTALL_DATAROOTDIR}/${PROJECT_NAME}/cmake
+)
+
+INSTALL(
+  DIRECTORY
+    ${PROJECT_SOURCE_DIR}/include/peelo
+  DESTINATION
+    include
+)
 
 ENABLE_TESTING()
 ADD_SUBDIRECTORY(test)
