--- src/CMakeLists.txt
@@ -1,18 +1,0 @@
-ADD_LIBRARY(
-  peelocpp_chrono
-  date.cpp
-  datetime.cpp
-  month.cpp
-  time.cpp
-  weekday.cpp
-)
-INSTALL(
-  DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/../include/peelo
-  DESTINATION include
-  FILES_MATCHING PATTERN "*.hpp"
-)
-INSTALL(
-  TARGETS peelocpp_chrono
-  ARCHIVE
-  DESTINATION lib
-)
