+++ test/test_month.cpp
@@ -1,27 +1,19 @@
 #include <peelo/chrono/month.hpp>
 #include <cassert>
-#include <sstream>
-
-static std::string month_to_string(const peelo::month& month)
-{
-  std::stringstream ss;
-
-  ss << month;
-
-  return ss.str();
-}
 
 int main()
 {
-  assert(static_cast<int>(peelo::month::jan) == 0);
+  using namespace peelo;
 
-  assert(peelo::month::jan + 3 == peelo::month::apr);
-  assert(peelo::month::sep - 2 == peelo::month::jul);
-  assert(peelo::month::jan + 13 == peelo::month::feb);
-  assert(peelo::month::nov - 13 == peelo::month::oct);
+  assert(static_cast<int>(chrono::month::jan) == 0);
 
-  assert(month_to_string(peelo::month::jan) == "January");
-  assert(month_to_string(peelo::month::oct) == "October");
+  assert(chrono::month::jan + 3 == chrono::month::apr);
+  assert(chrono::month::sep - 2 == chrono::month::jul);
+  assert(chrono::month::jan + 13 == chrono::month::feb);
+  assert(chrono::month::nov - 13 == chrono::month::oct);
+
+  assert(chrono::to_string(chrono::month::jan) == "January");
+  assert(chrono::to_string(chrono::month::oct) == "October");
 
   return 0;
 }
